+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add SDL_GetTicks64() (#461)
 * Implements snd_pcm_hw_params_get_rate()
 * Implements glXMakeContextCurrent()
+* Add a warning for possible time-tracking option
 
 ### Changed